modularized batch processing outside
